pyproject.toml
[tool.poetry]
name = "superheroes-azure-function"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-functions = "^1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
